pyproject.toml
[project]
name = "code-basics-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.31.1",
    "langchain-community>=0.3.29",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.7",
    "langgraph-supervisor>=0.0.29",
    "pygithub>=2.8.1",
    "python-dotenv>=1.1.1",
]
